20814: Migrate ablate & analyze operations (#229)
- Removes deprecated client load/unload methods
- Migrates analyze and ablation operations
- Adds new abstract method to check if tracing is enabled
- Some typing fixes
