+++ howso/direct/client.py
@@ -922,42 +922,6 @@ class HowsoDirectClient(AbstractHowsoClient):
             target_name_path=target_name_path
         )
 
-    def load_trainee(self, trainee_id: str):
-        """
-        Load a Trainee that was persisted on the Howso service.
-
-        .. deprecated:: 1.0.0
-            Use :meth:`HowsoDirectClient.acquire_trainee_resources` instead.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee load.
-        """
-        warnings.warn(
-            'The method `load_trainee()` is deprecated and will be removed in '
-            'a future release. Please use `acquire_trainee_resources()` '
-            'instead.', DeprecationWarning)
-        self.acquire_trainee_resources(trainee_id)
-
-    def unload_trainee(self, trainee_id: str):
-        """
-        Unload a Trainee from the Howso service.
-
-        .. deprecated:: 1.0.0
-            Use :meth:`HowsoDirectClient.release_trainee_resources` instead.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee unload.
-        """
-        warnings.warn(
-            'The method `unload_trainee()` is deprecated and will be removed '
-            'in a future release. Please use `release_trainee_resources()` '
-            'instead.', DeprecationWarning)
-        self.release_trainee_resources(trainee_id)
-
     def acquire_trainee_resources(
         self,
         trainee_id: str,
@@ -3941,364 +3905,6 @@ class HowsoDirectClient(AbstractHowsoClient):
             child_name_path=child_name_path
         )
 
-    def set_auto_analyze_params(  # noqa: C901
-        self,
-        trainee_id: str,
-        auto_analyze_enabled: bool = False,
-        analyze_threshold: Optional[int] = None,
-        *,
-        auto_analyze_limit_size: Optional[int] = None,
-        analyze_growth_factor: Optional[float] = None,
-        action_features: Optional[Iterable[str]] = None,
-        context_features: Optional[Iterable[str]] = None,
-        k_folds: Optional[int] = None,
-        num_samples: Optional[int] = None,
-        dt_values: Optional[Iterable[float]] = None,
-        k_values: Optional[Iterable[int]] = None,
-        p_values: Optional[Iterable[float]] = None,
-        bypass_hyperparameter_analysis: Optional[bool] = None,
-        bypass_calculate_feature_residuals: Optional[bool] = None,
-        bypass_calculate_feature_weights: Optional[bool] = None,
-        targeted_model: Optional[Literal["omni_targeted", "single_targeted", "targetless"]] = None,
-        num_analysis_samples: Optional[int] = None,
-        analysis_sub_model_size: Optional[int] = None,
-        use_deviations: Optional[bool] = None,
-        inverse_residuals_as_weights: Optional[bool] = None,
-        use_case_weights: Optional[bool] = None,
-        weight_feature: Optional[str] = None,
-        **kwargs
-    ):
-        """
-        Set trainee parameters for auto analysis.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee to set auto analysis parameters for.
-        auto_analyze_enabled : bool, default False
-            When True, the :func:`train` method will trigger an analyze when
-            it's time for the model to be analyzed again.
-        analyze_threshold : int, optional
-            The threshold for the number of cases at which the model should be
-            re-analyzed.
-        auto_analyze_limit_size : int, optional
-            The size of of the model at which to stop doing auto-analysis.
-            Value of 0 means no limit.
-        analyze_growth_factor : float, optional
-            The factor by which to increase the analyze threshold every time
-            the model grows to the current threshold size.
-        action_features : Iterable[str], optional
-            The action features to analyze for.
-        context_features : Iterable[str], optional
-            The context features to analyze for.
-        k_folds : int, optional
-            The number of cross validation folds to do. A value of 1 does
-            hold-one-out instead of k-fold.
-        num_samples : int, optional
-            Number of samples used in calculating feature residuals.
-        dt_values : Iterable[float], optional
-            The dt value hyperparameters to analyze with.
-        k_values : Iterable[int], optional
-            The number of cross validation folds to do. A value of 1 does
-            hold-one-out instead of k-fold.
-        p_values : Iterable[float], optional
-            The p value hyperparameters to analyze with.
-        bypass_calculate_feature_residuals : bool, optional
-            When True, bypasses calculation of feature residuals.
-        bypass_calculate_feature_weights : bool, optional
-            When True, bypasses calculation of feature weights.
-        bypass_hyperparameter_analysis : bool, optional
-            When True, bypasses hyperparameter analysis.
-        targeted_model : Literal["omni_targeted", "single_targeted", "targetless"], optional
-            Type of hyperparameter targeting.
-            Valid options include:
-
-                - **single_targeted**: Analyze hyperparameters for the
-                  specified action_features.
-                - **omni_targeted**: Analyze hyperparameters for each context
-                  feature as an action feature, ignores action_features
-                  parameter.
-                - **targetless**: Analyze hyperparameters for all context
-                  features as possible action features, ignores
-                  action_features parameter.
-        num_analysis_samples : int, optional
-            Specifies the number of observations to be considered for
-            analysis.
-        analysis_sub_model_size : int, optional
-            Number of samples to use for analysis. The rest will be
-            randomly held-out and not included in calculations.
-        use_deviations : bool, optional
-            When True, uses deviations for LK metric in queries.
-        inverse_residuals_as_weights : bool, optional
-            When True, will compute and use inverse of residuals as feature
-            weights.
-        use_case_weights : bool, optional
-            When True will scale influence weights by each
-            case's weight_feature weight.
-        weight_feature : str
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        kwargs : dict, optional
-            Parameters specific for analyze() may be passed in via kwargs, and
-            will be cached and used during future auto-analysis.
-        """
-        self._auto_resolve_trainee(trainee_id)
-
-        deprecated_params = {
-            'auto_optimize_enabled': 'auto_analyze_enabled',
-            'optimize_threshold': 'analyze_threshold',
-            'optimize_growth_factor': 'analyze_growth_factor',
-            'auto_optimize_limit_size': 'auto_analyze_limit_size',
-        }
-        analyze_deprecated_params = {
-            'bypass_hyperparameter_optimization': 'bypass_hyperparameter_analysis',
-            'num_optimization_samples': 'num_analysis_samples',
-            'optimization_sub_model_size': 'analysis_sub_model_size',
-            'dwe_values': 'dt_values'
-        }
-
-        # explicitly update parameters if old names are provided
-        if kwargs:
-            for old_param, new_param in deprecated_params.items():
-                if old_param in kwargs:
-                    if old_param == 'auto_optimize_enabled':
-                        auto_analyze_enabled = kwargs[old_param]
-                    elif old_param == 'optimize_threshold':
-                        analyze_threshold = kwargs[old_param]
-                    elif old_param == 'optimize_growth_factor':
-                        analyze_growth_factor = kwargs[old_param]
-                    elif old_param == 'auto_optimize_limit_size':
-                        auto_analyze_limit_size = kwargs[old_param]
-
-                    del kwargs[old_param]
-                    warnings.warn(
-                        f'The `{old_param}` parameter has been renamed to '
-                        f'`{new_param}`, please use the new parameter '
-                        'instead.', UserWarning)
-
-            # replace any old kwarg param with new param and remove old param
-            for old_param, new_param in analyze_deprecated_params.items():
-                if old_param in kwargs:
-                    kwargs[new_param] = kwargs[old_param]
-                    del kwargs[old_param]
-                    warnings.warn(
-                        f'The `{old_param}` parameter has been renamed to '
-                        f'`{new_param}`, please use the new parameter '
-                        'instead.', UserWarning)
-
-        if 'targeted_model' in kwargs:
-            targeted_model = kwargs['targeted_model']
-            if targeted_model not in ['single_targeted', 'omni_targeted', 'targetless']:
-                raise ValueError(
-                    f'Invalid value "{targeted_model}" for targeted_model. '
-                    'Valid values include single_targeted, omni_targeted, '
-                    'and targetless.')
-
-        # Collect valid parameters
-        if kwargs:
-            warn_params = ', '.join(kwargs)
-            warnings.warn(
-                f'The following auto ablation parameter(s) "{warn_params}" '
-                'are not officially supported or may not have an effect.', UserWarning)
-
-        self.howso.auto_analyze_params(
-            trainee_id=trainee_id,
-            auto_analyze_enabled=auto_analyze_enabled,
-            analyze_threshold=analyze_threshold,
-            auto_analyze_limit_size=auto_analyze_limit_size,
-            analyze_growth_factor=analyze_growth_factor,
-            action_features=action_features,
-            context_features=context_features,
-            k_folds=k_folds,
-            num_samples=num_samples,
-            dt_values=dt_values,
-            k_values=k_values,
-            p_values=p_values,
-            bypass_hyperparameter_analysis=bypass_hyperparameter_analysis,
-            bypass_calculate_feature_residuals=bypass_calculate_feature_residuals,
-            bypass_calculate_feature_weights=bypass_calculate_feature_weights,
-            targeted_model=targeted_model,
-            num_analysis_samples=num_analysis_samples,
-            analysis_sub_model_size=analysis_sub_model_size,
-            use_deviations=use_deviations,
-            inverse_residuals_as_weights=inverse_residuals_as_weights,
-            use_case_weights=use_case_weights,
-            weight_feature=weight_feature,
-            **kwargs
-        )
-        self._auto_persist_trainee(trainee_id)
-
-    def get_auto_ablation_params(self, trainee_id: str):
-        """Get trainee parameters for auto-ablation set by :meth:`set_auto_ablation_params`."""
-        self._auto_resolve_trainee(trainee_id)
-        return self.howso.get_auto_ablation_params(trainee_id)
-
-    def set_auto_ablation_params(
-        self,
-        trainee_id: str,
-        auto_ablation_enabled: bool = False,
-        *,
-        auto_ablation_weight_feature: str = ".case_weight",
-        conviction_lower_threshold: Optional[float] = None,
-        conviction_upper_threshold: Optional[float] = None,
-        exact_prediction_features: Optional[List[str]] = None,
-        influence_weight_entropy_threshold: float = 0.6,
-        minimum_model_size: int = 1_000,
-        relative_prediction_threshold_map: Optional[Dict[str, float]] = None,
-        residual_prediction_features: Optional[List[str]] = None,
-        tolerance_prediction_threshold_map: Optional[Dict[str, Tuple[float, float]]] = None,
-        **kwargs
-    ):
-        """
-        Set trainee parameters for auto-ablation.
-
-        .. note::
-            All ablation endpoints, including :meth:`set_auto_ablation_params` are experimental and may
-            have their API changed without deprecation.
-
-        .. seealso::
-            The params ``influence_weight_entropy_threshold`` and ``auto_ablation_weight_feature`` that are
-            set using this endpoint are used as defaults by :meth:`reduce_data`.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee to set auto ablation parameters for.
-        auto_ablation_enabled : bool, default False
-            When True, the :meth:`train` method will ablate cases that meet the set criteria.
-        auto_ablation_weight_feature : str, default ".case_weight"
-            The weight feature that should be accumulated to when cases are ablated.
-        minimum_model_size : int, default 1,000
-            The threshold of the minimum number of cases at which the model should auto-ablate.
-        influence_weight_entropy_threshold : float, default 0.6
-            The influence weight entropy quantile that a case must be beneath in order to be trained.
-        exact_prediction_features : Optional[List[str]], optional
-            For each of the features specified, will ablate a case if the prediction matches exactly.
-        residual_prediction_features : Optional[List[str]], optional
-            For each of the features specified, will ablate a case if
-            abs(prediction - case value) / prediction <= feature residual.
-        tolerance_prediction_threshold_map : Optional[Dict[str, Tuple[float, float]]], optional
-            For each of the features specified, will ablate a case if the prediction >= (case value - MIN)
-            and the prediction <= (case value + MAX).
-        relative_prediction_threshold_map : Optional[Dict[str, float]], optional
-            For each of the features specified, will ablate a case if
-            abs(prediction - case value) / prediction <= relative threshold
-        conviction_lower_threshold : Optional[float], optional
-            The conviction value above which cases will be ablated.
-        conviction_upper_threshold : Optional[float], optional
-            The conviction value below which cases will be ablated.
-        """
-        params = dict(
-            auto_ablation_enabled=auto_ablation_enabled,
-            auto_ablation_weight_feature=auto_ablation_weight_feature,
-            minimum_model_size=minimum_model_size,
-            influence_weight_entropy_threshold=influence_weight_entropy_threshold,
-            exact_prediction_features=exact_prediction_features,
-            residual_prediction_features=residual_prediction_features,
-            tolerance_prediction_threshold_map=tolerance_prediction_threshold_map,
-            relative_prediction_threshold_map=relative_prediction_threshold_map,
-            conviction_lower_threshold=conviction_lower_threshold,
-            conviction_upper_threshold=conviction_upper_threshold,
-        )
-        params.update(kwargs)
-        if kwargs:
-            warn_params = ", ".join(kwargs)
-            warnings.warn(
-                f"The following parameter(s) are not officially supported by `reduce_data` and "
-                f"may or may not have an effect: {warn_params}",
-                UnsupportedArgumentWarning
-            )
-        self._auto_resolve_trainee(trainee_id)
-        self.howso.set_auto_ablation_params(
-            trainee_id, **params
-        )
-
-    def reduce_data(
-        self,
-        trainee_id: str,
-        features: t.Optional[list[str]] = None,
-        distribute_weight_feature: t.Optional[str] = None,
-        influence_weight_entropy_threshold: t.Optional[float] = None,
-        skip_auto_analyze: bool = False,
-        **kwargs,
-    ):
-        """
-        Smartly reduce the amount of trained cases while accumulating case weights.
-
-        Determines which cases to remove by comparing the influence weight entropy of each trained
-        case to the ``influence_weight_entropy_threshold`` quantile of existing influence weight
-        entropies.
-
-        .. note::
-            All ablation endpoints, including :meth:`reduce_data` are experimental and may have their
-            API changed without deprecation.
-
-        .. seealso::
-            The default ``distribute_weight_feature`` and ``influence_weight_entropy_threshold`` are
-            pulled from the auto-ablation parameters, which can be set or retrieved with
-            :meth:`set_auto_ablation_params` and :meth:`get_auto_ablation_params`, respectively.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee for which to reduce data.
-        features : list of str, optional
-            The features which should be used to determine which cases to remove. This defaults to all of
-            the trained features (excluding internal features).
-        distribute_weight_feature : str, optional
-            The name of the weight feature to accumulate case weights to as cases are removed. This
-            defaults to the value of ``auto_ablation_weight_feature`` from :meth:`set_auto_ablation_params`,
-            which defaults to ".case_weight".
-        influence_weight_entropy_threshold : float, optional
-            The quantile of influence weight entropy above which cases will be removed. This defaults
-            to the value of ``influence_weight_entropy_threshold`` from :meth:`set_auto_ablation_params`,
-            which defaults to 0.6.
-        skip_auto_analyze : bool, default False
-            Whether to skip auto-analyzing as cases are removed.
-        """
-        params = dict(
-            features=features,
-            distribute_weight_feature=distribute_weight_feature,
-            influence_weight_entropy_threshold=influence_weight_entropy_threshold,
-            skip_auto_analyze=skip_auto_analyze,
-        )
-
-        params.update(kwargs)
-        if kwargs:
-            warn_params = ", ".join(kwargs)
-            warnings.warn(
-                f"The following parameter(s) are not officially supported by `reduce_data` and "
-                f"may or may not have an effect: {warn_params}",
-                UnsupportedArgumentWarning
-            )
-        self._auto_resolve_trainee(trainee_id)
-        self.howso.reduce_data(trainee_id, **params)
-
-    def auto_analyze(self, trainee_id: str):
-        """
-        Auto-analyze the trainee model.
-
-        Re-uses all parameters from the previous analyze or
-        set_auto_analyze_params call. If analyze or set_auto_analyze_params
-        has not been previously called, auto_analyze will default to a robust
-        and versatile analysis.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee to auto-analyze.
-        """
-        self._auto_resolve_trainee(trainee_id)
-        if self.verbose:
-            print(f"Auto-analyzing trainee with id: {trainee_id}")
-
-        self.howso.auto_analyze(trainee_id)
-        self._auto_persist_trainee(trainee_id)
-        # when debugging output the auto-analyzed parameters into the
-        # trace file.
-        if self.howso.trace:
-            self.howso.get_internal_parameters(trainee_id)
-
     def get_label(self, entity_id: str, label: str) -> object:
         """
         Get a label value from a Trainee.
@@ -4400,162 +4006,3 @@ class HowsoDirectClient(AbstractHowsoClient):
             weight_feature, use_case_weights)
         self._auto_persist_trainee(trainee_id)
         return weights
-
-    def analyze(
-        self,
-        trainee_id: str,
-        context_features: Optional[Iterable[str]] = None,
-        action_features: Optional[Iterable[str]] = None,
-        *,
-        bypass_calculate_feature_residuals: bool = None,
-        bypass_calculate_feature_weights: bool = None,
-        bypass_hyperparameter_analysis: bool = None,
-        dt_values: Optional[List[float]] = None,
-        use_case_weights: bool = None,
-        inverse_residuals_as_weights: bool = None,
-        k_folds: Optional[int] = None,
-        k_values: Optional[List[int]] = None,
-        num_analysis_samples: Optional[int] = None,
-        num_samples: Optional[int] = None,
-        analysis_sub_model_size: Optional[int] = None,
-        p_values: Optional[List[float]] = None,
-        targeted_model: Optional[Literal["omni_targeted", "single_targeted", "targetless"]] = None,
-        use_deviations: bool = None,
-        weight_feature: Optional[str] = None,
-        **kwargs
-    ):
-        """
-        Analyzes a trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee.
-        context_features : iterable of str, optional
-            The context features to analyze for.
-        action_features : iterable of str, optional
-            The action features to analyze for.
-        k_folds : int
-            optional, (default 6) number of cross validation folds to do
-        bypass_hyperparameter_analysis : bool
-            optional, bypasses hyperparameter analysis
-        bypass_calculate_feature_residuals : bool
-            optional, bypasses feature residual calculation
-        bypass_calculate_feature_weights : bool
-            optional, bypasses calculation of feature weights
-        use_deviations : bool
-            optional, uses deviations for LK metric in queries
-        num_samples : int
-            used in calculating feature residuals
-        k_values : list of int
-            optional list used in hyperparameter search
-        p_values : list of float
-            optional list used in hyperparameter search
-        dt_values : list of float
-            optional list used in hyperparameter search
-        targeted_model : {"omni_targeted", "single_targeted", "targetless"}
-            optional, valid values as follows:
-
-                "single_targeted" = analyze hyperparameters for the
-                    specified action_features
-                "omni_targeted" = analyze hyperparameters for each context
-                    feature as an action feature, ignores action_features
-                    parameter
-                "targetless" = analyze hyperparameters for all context
-                    features as possible action features, ignores
-                    action_features parameter
-        num_analysis_samples : int, optional
-            If the dataset size to too large, analyze on
-            (randomly sampled) subset of data. The
-            `num_analysis_samples` specifies the number of
-            observations to be considered for analysis.
-        analysis_sub_model_size : int or Node, optional
-            Number of samples to use for analysis. The rest
-            will be randomly held-out and not included in calculations.
-        inverse_residuals_as_weights : bool, default is False
-            When True will compute and use inverse of residuals
-            as feature weights
-        use_case_weights : bool, default False
-            When True will scale influence weights by each
-            case's weight_feature weight.
-        weight_feature : str, optional
-            Name of feature whose values to use as case weights.
-            When left unspecified uses the internally managed case weight.
-        kwargs
-            Additional experimental analyze parameters.
-        """
-        self._auto_resolve_trainee(trainee_id)
-
-        validate_list_shape(context_features, 1, "context_features", "str")
-        validate_list_shape(action_features, 1, "action_features", "str")
-        validate_list_shape(p_values, 1, "p_values", "int")
-        validate_list_shape(k_values, 1, "k_values", "float")
-        validate_list_shape(dt_values, 1, "dt_values", "float")
-
-        if targeted_model not in ['single_targeted', 'omni_targeted', 'targetless', None]:
-            raise ValueError(
-                f'Invalid value "{targeted_model}" for targeted_model. '
-                'Valid values include single_targeted, omni_targeted, '
-                'and targetless.')
-
-        deprecated_params = {
-            'bypass_hyperparameter_optimization': 'bypass_hyperparameter_analysis',
-            'num_optimization_samples': 'num_analysis_samples',
-            'optimization_sub_model_size': 'analysis_sub_model_size',
-            'dwe_values': 'dt_values'
-        }
-        # explicitly update parameters if old names are provided
-        if kwargs:
-            for old_param, new_param in deprecated_params.items():
-                if old_param in kwargs:
-                    if old_param == 'bypass_hyperparameter_optimization':
-                        bypass_hyperparameter_analysis = kwargs[old_param]
-                    elif old_param == 'num_optimization_samples':
-                        num_analysis_samples = kwargs[old_param]
-                    elif old_param == 'optimization_sub_model_size':
-                        analysis_sub_model_size = kwargs[old_param]
-                    elif old_param == 'dwe_values':
-                        dt_values = kwargs[old_param]
-
-                    del kwargs[old_param]
-                    warnings.warn(
-                        f'The `{old_param}` parameter has been renamed to '
-                        f'`{new_param}`, please use the new parameter '
-                        'instead.', UserWarning)
-
-        analyze_params = dict(
-            action_features=action_features,
-            context_features=context_features,
-            bypass_calculate_feature_residuals=bypass_calculate_feature_residuals,  # noqa: #E501
-            bypass_calculate_feature_weights=bypass_calculate_feature_weights,
-            bypass_hyperparameter_analysis=bypass_hyperparameter_analysis,  # noqa: #E501
-            dt_values=dt_values,
-            use_case_weights=use_case_weights,
-            inverse_residuals_as_weights=inverse_residuals_as_weights,
-            k_folds=k_folds,
-            k_values=k_values,
-            num_analysis_samples=num_analysis_samples,
-            num_samples=num_samples,
-            analysis_sub_model_size=analysis_sub_model_size,
-            p_values=p_values,
-            targeted_model=targeted_model,
-            use_deviations=use_deviations,
-            weight_feature=weight_feature,
-        )
-
-        # Add experimental options
-        analyze_params.update(kwargs)
-
-        if kwargs:
-            warn_params = ', '.join(kwargs)
-            warnings.warn(
-                f'The following analyze parameter(s) "{warn_params}" are '
-                'not officially supported by analyze and may or may not '
-                'have an effect.', UserWarning)
-
-        if self.verbose:
-            print(f'Analyzing trainee with id: {trainee_id}')
-            print(f'Analyzing trainee with parameters: {analyze_params}')
-
-        self.howso.analyze(trainee_id, **analyze_params)
-        self._auto_persist_trainee(trainee_id)
