+++ howso/direct/core.py
@@ -724,187 +724,6 @@ class HowsoCore:
             }
         )
 
-    def get_auto_ablation_params(self, trainee_id: str):
-        """Get trainee parameters for auto-ablation set by :meth:`set_auto_ablation_params`."""
-        return self._execute(
-            trainee_id, "get_auto_ablation_params", {}
-        )
-
-    def set_auto_ablation_params(
-        self,
-        trainee_id: str,
-        auto_ablation_enabled: bool = False,
-        *,
-        auto_ablation_weight_feature: str = ".case_weight",
-        conviction_lower_threshold: Optional[float] = None,
-        conviction_upper_threshold: Optional[float] = None,
-        exact_prediction_features: Optional[List[str]] = None,
-        influence_weight_entropy_threshold: float = 0.6,
-        minimum_model_size: int = 1_000,
-        relative_prediction_threshold_map: Optional[Dict[str, float]] = None,
-        residual_prediction_features: Optional[List[str]] = None,
-        tolerance_prediction_threshold_map: Optional[Dict[str, Tuple[float, float]]] = None,
-        **kwargs
-    ):
-        """
-        Set trainee parameters for auto-ablation.
-
-        .. note::
-            All ablation endpoints, including :meth:`set_auto_ablation_params` are experimental and may
-            have their API changed without deprecation.
-
-        .. seealso::
-            The params ``influence_weight_entropy_threshold`` and ``auto_ablation_weight_feature`` that are
-            set using this endpoint are used as defaults by :meth:`reduce_data`.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee to set auto ablation parameters for.
-        auto_ablation_enabled : bool, default False
-            When True, the :meth:`train` method will ablate cases that meet the set criteria.
-        auto_ablation_weight_feature : str, default ".case_weight"
-            The weight feature that should be accumulated to when cases are ablated.
-        minimum_model_size : int, default 1,000
-            The threshold ofr the minimum number of cases at which the model should auto-ablate.
-        influence_weight_entropy_threshold : float, default 0.6
-            The influence weight entropy quantile that a case must be beneath in order to be trained.
-        exact_prediction_features : Optional[List[str]], optional
-            For each of the features specified, will ablate a case if the prediction matches exactly.
-        residual_prediction_features : Optional[List[str]], optional
-            For each of the features specified, will ablate a case if
-            abs(prediction - case value) / prediction <= feature residual.
-        tolerance_prediction_threshold_map : Optional[Dict[str, Tuple[float, float]]], optional
-            For each of the features specified, will ablate a case if the prediction >= (case value - MIN)
-            and the prediction <= (case value + MAX).
-        relative_prediction_threshold_map : Optional[Dict[str, float]], optional
-            For each of the features specified, will ablate a case if
-            abs(prediction - case value) / prediction <= relative threshold
-        conviction_lower_threshold : Optional[float], optional
-            The conviction value above which cases will be ablated.
-        conviction_upper_threshold : Optional[float], optional
-            The conviction value below which cases will be ablated.
-        """
-        return self._execute(
-            trainee_id, "set_auto_ablation_params",
-            {
-                "auto_ablation_enabled": auto_ablation_enabled,
-                "auto_ablation_weight_feature": auto_ablation_weight_feature,
-                "minimum_model_size": minimum_model_size,
-                "influence_weight_entropy_threshold": influence_weight_entropy_threshold,
-                "exact_prediction_features": exact_prediction_features,
-                "residual_prediction_features": residual_prediction_features,
-                "tolerance_prediction_threshold_map": tolerance_prediction_threshold_map,
-                "relative_prediction_threshold_map": relative_prediction_threshold_map,
-                "conviction_lower_threshold": conviction_lower_threshold,
-                "conviction_upper_threshold": conviction_upper_threshold,
-            }
-        )
-
-    def reduce_data(
-        self,
-        trainee_id: str,
-        features: t.Optional[list[str]] = None,
-        distribute_weight_feature: t.Optional[str] = None,
-        influence_weight_entropy_threshold: t.Optional[float] = None,
-        skip_auto_analyze: bool = False,
-        **kwargs,
-    ):
-        """
-        Smartly reduce the amount of trained cases while accumulating case weights.
-
-        Determines which cases to remove by comparing the influence weight entropy of each trained
-        case to the ``influence_weight_entropy_threshold`` quantile of existing influence weight
-        entropies.
-
-        .. note::
-            All ablation endpoints, including :meth:`reduce_data` are experimental and may have their
-            API changed without deprecation.
-
-        .. seealso::
-            The default ``distribute_weight_feature`` and ``influence_weight_entropy_threshold`` are
-            pulled from the auto-ablation parameters, which can be set or retrieved with
-            :meth:`set_auto_ablation_params` and :meth:`get_auto_ablation_params`, respectively.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The ID of the Trainee for which to reduce data.
-        features : list of str, optional
-            The features which should be used to determine which cases to remove. This defaults to all of
-            the trained features (excluding internal features).
-        distribute_weight_feature : str, optional
-            The name of the weight feature to accumulate case weights to as cases are removed. This
-            defaults to the value of ``auto_ablation_weight_feature`` from :meth:`set_auto_ablation_params`,
-            which defaults to ".case_weight".
-        influence_weight_entropy_threshold : float, optional
-            The quantile of influence weight entropy above which cases will be removed. This defaults
-            to the value of ``influence_weight_entropy_threshold`` from :meth:`set_auto_ablation_params`,
-            which defaults to 0.6.
-        skip_auto_analyze : bool, default False
-            Whether to skip auto-analyzing as cases are removed.
-        """
-        return self._execute(
-            trainee_id, "reduce_data",
-            {
-                "features": features,
-                "distribute_weight_feature": distribute_weight_feature,
-                "influence_weight_entropy_threshold": influence_weight_entropy_threshold,
-                "skip_auto_analyze": skip_auto_analyze,
-            }
-        )
-
-    def auto_analyze_params(
-        self,
-        trainee_id: str,
-        auto_analyze_enabled: bool = False,
-        analyze_threshold: Optional[int] = None,
-        auto_analyze_limit_size: Optional[int] = None,
-        analyze_growth_factor: Optional[float] = None,
-        **kwargs
-    ) -> None:
-        """
-        Set trainee parameters for auto analysis.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        auto_analyze_enabled : bool, default False
-            Enable auto analyze when training. Train will return a status
-            indicating when to auto analyze.
-        analyze_threshold : int, optional
-            The threshold for the number of cases at which the model should be
-            re-analyzed.
-        auto_analyze_limit_size : int, optional
-            The size of of the model at which to stop doing auto-analysis.
-            Value of 0 means no limit.
-        analyze_growth_factor : float, optional
-            The factor by which to increase the analyze threshold every time
-            the model grows to the current threshold size.
-        kwargs : dict, optional
-            Parameters specific for analyze() may be passed in via kwargs, and
-            will be cached and used during future auto-analysis.
-        """
-        params = {
-            "auto_analyze_enabled": auto_analyze_enabled,
-            "analyze_threshold": analyze_threshold,
-            "analyze_growth_factor": analyze_growth_factor,
-            "auto_analyze_limit_size": auto_analyze_limit_size,
-        }
-        return self._execute(trainee_id, "set_auto_analyze_params", {**kwargs, **params})
-
-    def auto_analyze(self, trainee_id: str) -> None:
-        """
-        Auto-analyze the Trainee model.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        """
-        return self._execute(trainee_id, "auto_analyze", {})
-
     def compute_feature_weights(
         self,
         trainee_id: str,
@@ -2166,20 +1985,6 @@ class HowsoCore:
             "root_filepath": f"{self.howso_path}/",
             "separate_files": separate_files,
         })
-
-    def analyze(self, trainee_id: str, **kwargs) -> None:
-        """
-        Analyzes a trainee.
-
-        Parameters
-        ----------
-        trainee_id : str
-            The identifier of the Trainee.
-        kwargs : dict
-            Analysis arguments.
-        """
-        params = {**kwargs}
-        return self._execute(trainee_id, "analyze", params)
 
     def reset_parameter_defaults(self, trainee_id: str) -> None:
         """
